leakage_agent/pyproject.toml
[project]
name = "leakage_agent"
version = "0.1.0"
description = "Data Leakage Auto-Supervisor Agent"
dependencies = [
    "pandas",
    "pyyaml",
    "python-dateutil",
    "jsonschema"
]

[project.scripts]
leakage-agent = "leakage_agent.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
